requirements.txt
asgiref==3.8.1
django-cors-headers==4.7.0
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
mysqlclient==2.2.7
PyMySQL==1.1.1
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2